ui: split 'VUnit' to ui/vunit.py
